api/requirements.txt
# general
python-dateutil
requests

# app
fastapi
fastui
python-multipart
uvicorn[standard]

# data
Jinja2
psycopg2

# tests
pytest